+++ frontend/src/context/AuthProvider.tsx
@@ -1,18 +1,13 @@
-import { createContext, ReactNode, useContext, useState } from "react";
+import { ReactNode, useState } from "react";
+import { AuthContext } from "./AuthContext";
 
-// Define authentication context
-interface AuthContextType {
-  user: string | null;
-  login: (username: string) => void;
-  logout: () => void;
-}
-
-const AuthContext = createContext<AuthContextType | undefined>(undefined);
-
-const AuthProvider = ({ children }: { children: ReactNode }) => {
+export const AuthProvider = ({ children }: { children: ReactNode }) => {
   const [user, setUser] = useState<string | null>(null);
 
-  const login = (username: string) => setUser(username);
+  const login = (username: string) => {
+    console.log("Setting user")
+    setUser(username)
+  };
   const logout = () => setUser(null);
 
   return (
@@ -21,14 +16,3 @@ const AuthProvider = ({ children }: { children: ReactNode }) => {
     </AuthContext.Provider>
   );
 };
-
-const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (!context) {
-    throw new Error("useAuth must be used within an AuthProvider");
-  }
-  return context;
-};
-
-export { AuthProvider, useAuth };
-
